Tests/LibWeb/Text/expected/element-get-bounding-client-rect-nested-transform-origins.txt

x close to 50: false
y close to 150: false
width: 50
height: 50
